Move fixture to plotsearch conftest
